Backend/requirement.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
requests>=2.26.0
geopy>=2.2.0
flask>=2.0.0
python-dotenv>=0.19.0
pytest>=6.2.0
jupyter>=1.0.0
matplotlib
langgraph
langchain
langchain-core
langchain-community
langchain-groq
pygraphviz
fastapi
uvicorn